% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.tokens_list}
\alias{as.data.frame.tokens_list}
\alias{interopertability}
\alias{as.list.tokens}
\title{Interoperability with other text analysis packages}
\usage{
\method{as.data.frame}{tokens_list}(x)

\method{as.list}{tokens}(x)
}
\description{
These functions are provided for interoperability with other R
  text analysis packages. They convert a named list of tokens to a data.frame
  that meets the specifications for interopertability, or vice versa.
}
\examples{
tokens_l <- tokenize_words(c(a = "A short text.", b = "Even shorter."))
as.data.frame(tokens_l)
as.list(as.data.frame(tokens_l))
}
