% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizers-package.r
\docType{package}
\name{tokenizers}
\alias{tokenizers-package}
\alias{tokenizers}
\title{Tokenizers}
\description{
A collection of functions with a consistent interface to convert natural
language text into tokens.
}
\details{
The tokenizers in this package have a consistent interface. They all take
either a character vector of any length, or a list where each element is a
character vector of length one. The idea is that each element comprises a
text. Then each function returns a list with the same length as the input
vector, where each element in the list are the tokens generated by the
function. If the input character vector or list is named, then the names are
preserved.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/tokenizers/}
  \item \url{https://github.com/ropensci/tokenizers}
  \item Report bugs at \url{https://github.com/ropensci/tokenizers/issues}
}

}
\author{
\strong{Maintainer}: Thomas Charlon \email{charlon@protonmail.com} (\href{https://orcid.org/0000-0001-7497-0470}{ORCID})

Authors:
\itemize{
  \item Lincoln Mullen \email{lincoln@lincolnmullen.com} (\href{https://orcid.org/0000-0001-5103-6917}{ORCID})
}

Other contributors:
\itemize{
  \item Os Keyes \email{ironholds@gmail.com} (\href{https://orcid.org/0000-0001-5196-609X}{ORCID}) [contributor]
  \item Dmitriy Selivanov \email{selivanov.dmitriy@gmail.com} [contributor]
  \item Jeffrey Arnold \email{jeffrey.arnold@gmail.com} (\href{https://orcid.org/0000-0001-9953-3904}{ORCID}) [contributor]
  \item Kenneth Benoit \email{kbenoit@lse.ac.uk} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID}) [contributor]
}

}
