% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tokens_list_to_df}
\alias{tokens_list_to_df}
\alias{tokens_df_to_list}
\title{Interoperability with other text analysis packages}
\usage{
tokens_list_to_df(x)

tokens_df_to_list(x)
}
\arguments{
\item{x}{A list or data frame containing tokens from this package.}
}
\description{
These functions are provided for interoperability with other R
  text analysis packages. They convert a named list of tokens to a data.frame
  that meets the specifications for interopertability, or vice versa.
}
\examples{
tokens_l <- tokenize_words(c(a = "A short text.", b = "Even shorter."))
tokens_list_to_df(tokens_l)
tokens_df_to_list(tokens_list_to_df(tokens_l))
}
