% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{mobydick}
\alias{mobydick}
\title{The text of Moby Dick}
\format{A named character vector with length 1.}
\source{
\url{http://www.gutenberg.org/}
}
\usage{
mobydick
}
\description{
The text of Moby Dick, by Herman Melville, taken from Project Gutenberg.
}
\keyword{datasets}
