% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{stopwords}
\alias{stopwords}
\title{Stopword lists
Retrieve lists of stopwords by language.}
\usage{
stopwords(language = c("en", "da", "de", "el", "es", "fr", "it"))
}
\arguments{
\item{language}{The two-letter code for the name of the language.}
}
\description{
Stopword lists
Retrieve lists of stopwords by language.
}
\examples{
stopwords("en")
stopwords("de")
}
\references{
The stopword lists are a subset of the stopword lists available in the \href{Apache Lucene/Solr}{https://github.com/apache/lucene-solr/tree/master/solr/contrib/morphlines-core/src/test-files/solr/collection1/conf/lang} project, available under the Apache 2.0 license.
}

