% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizers-package.r
\docType{package}
\name{tokenizers}
\alias{tokenizers}
\alias{tokenizers-package}
\title{Tokenizers}
\description{
A collection of functions with a consistent interface to convert natural
language text into tokens.
}

